pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carm-company-profile"
version = "0.1.0"
description = "Company profile management with auto local/cloud storage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Justin Carm", email = "justin@carmvisuals.com"}
]
keywords = ["profile", "company", "storage", "azure", "cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "carm-data-models>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]

[project.urls]
Homepage = "https://github.com/JustinCarm001/carm-company-profile"
Repository = "https://github.com/JustinCarm001/carm-company-profile"
Issues = "https://github.com/JustinCarm001/carm-company-profile/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
carm_company_profile = ["profiles/*.json"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=carm_company_profile --cov-report=term-missing"
